[refactor] :truck: move signers balance checker to blockchain-service
