+++ src/services/transferService.ts
@@ -8,7 +8,6 @@ import { TokenBalanceType } from '../types/common';
 import Transaction from '../models/transaction';
 import { getEntryPointABI } from './bucketService';
 import { IBlockchain } from '../models/blockchain';
-import { getBlockchain } from './blockchainService';
 import { generatePrivateKey } from '../utils/keyGenerator';
 import { sendUserOperationToBundler } from './bundlerService';
 import { waitForUserOperationReceipt } from '../utils/waitForTX';
@@ -22,10 +21,10 @@ import {
   createGenericUserOperation
 } from './userOperationService';
 import {
-  USER_SIGNER_MIN_BALANCE,
-  BACKEND_SIGNER_MIN_BALANCE,
-  USER_SIGNER_BALANCE_TO_TRANSFER
-} from '../constants/environment';
+  getBlockchain,
+  ensureUserSignerHasEnoughEth,
+  ensureBackendSignerHasEnoughEth
+} from './blockchainService';
 
 /**
  * Sends a user operation for token transfer.
@@ -157,87 +156,6 @@ export async function sendUserOperation(
     );
     Logger.log('Full error object:', JSON.stringify(error));
     throw error;
-  }
-}
-
-/**
- * Helper function to ensure the User signer has enough ETH for gas fees.
- */
-export async function ensureUserSignerHasEnoughEth(
-  userSignerWalletAdddress: string,
-  backendSignerWallet: ethers.Wallet,
-  provider: ethers.providers.JsonRpcProvider
-): Promise<boolean> {
-  try {
-    const backendSignerWalletAddress = await backendSignerWallet.getAddress();
-    Logger.log(
-      `Checking if User EOA ${userSignerWalletAdddress} has minimal funds (${USER_SIGNER_MIN_BALANCE}) to sign the transaction.`
-    );
-
-    const EOABalance = await provider.getBalance(userSignerWalletAdddress);
-    Logger.log(
-      `User EOA ${userSignerWalletAdddress} balance: ${ethers.utils.formatEther(EOABalance)} ETH.`
-    );
-
-    if (EOABalance.lt(ethers.utils.parseEther(USER_SIGNER_MIN_BALANCE))) {
-      Logger.log(
-        `Sending ${USER_SIGNER_BALANCE_TO_TRANSFER} ETH from backendSigner ${backendSignerWalletAddress} ` +
-          `to User EOA ${userSignerWalletAdddress}.`
-      );
-
-      const tx = await backendSignerWallet.sendTransaction({
-        to: userSignerWalletAdddress,
-        value: ethers.utils.parseEther(USER_SIGNER_BALANCE_TO_TRANSFER),
-        gasLimit: 210000
-      });
-
-      await tx.wait();
-      Logger.log('ETH sent to user EOA');
-    } else {
-      Logger.log('User EOA has enough ETH to sign.');
-    }
-    return true;
-  } catch (error: unknown) {
-    Logger.error(
-      `Error checking if EOA Signer has minimal funds to make the transaction. Error: ${(error as Error).message}`
-    );
-    return false;
-  }
-}
-
-/**
- * Helper function to ensure the backend Signer has enough ETH for gas fees.
- */
-export async function ensureBackendSignerHasEnoughEth(
-  backendSignerWalletAddress: string,
-  provider: ethers.providers.JsonRpcProvider
-): Promise<boolean> {
-  try {
-    Logger.log(
-      `Checking if Backend Signer ${backendSignerWalletAddress} has minimal funds (${BACKEND_SIGNER_MIN_BALANCE}) to make transactions.`
-    );
-
-    const walletBalance = await provider.getBalance(backendSignerWalletAddress);
-    const walletBalanceFormatted = ethers.utils.formatEther(walletBalance);
-    Logger.log(
-      `Backend Signer ${backendSignerWalletAddress} balance: ${walletBalanceFormatted} ETH.`
-    );
-
-    if (walletBalance.lt(ethers.utils.parseEther(BACKEND_SIGNER_MIN_BALANCE))) {
-      Logger.error(
-        `Backend Signer ${backendSignerWalletAddress} current balance: ${walletBalanceFormatted} ETH, ` +
-          `balance required: ${BACKEND_SIGNER_MIN_BALANCE} ETH.`
-      );
-      return false;
-    }
-
-    Logger.log('Backend Signer has enough ETH.');
-    return true;
-  } catch (error: unknown) {
-    Logger.error(
-      `Error checking if Backend Signer has minimal funds to make transactions. Error: ${(error as Error).message}`
-    );
-    return false;
   }
 }
